+++ src/lib.rs
@@ -8,12 +8,15 @@
   A parser and evaluator of spreadsheet-like formulas
 </h3>
 
-It's in its early stages, and we are trying to add more functions and features soon.
+Formula is in its early stages and is not ready for production use.
 
 So far we have the following features:
 
 - 18 date time functions
 - 26 text functions
+- 7 logical functions
+- 2 web functions
+- plus all arithmetic and comparison operators
 
 ## Installation and usage
 
@@ -26,18 +29,19 @@ use formula::{Formula, Expr, error::Error};
 use anyhow::Result;
 
 fn main() -> Result<()> {
-    let formula = Formula::new("UPPER(TRIM('   Hello '))")?;
-    let value = formula.parse().unwrap();
+    let formula = Formula::new("=UPPER(TRIM('   Hello '))")?;
+    let value = formula.parse()?;
     assert_eq!(value, Expr::String("HELLO".to_string()));
     Ok(())
 }
 ```
 
-## What we do not support:
+## What we do not support, yet:
 
-- We would like to add more functions, like Excel functions, Google Sheets functions, and more
+- Support of functions are so limited at the moment, but we would like to add more of them, like Excel functions, Google Sheets functions, and so on
 - At the moment, we don't support table data, so you need to pass values to the formula as arguments by yourself
-- We do not support simple formulas like `1+1` or as argument like `AND(1>3, 1<3)` or `SUM(2-1, 2)`, yet
+- We do not support simple formulas like `1+1` or as argument like `AND(1>3, 1<3)` or `SUM(2-1, 2)`. Instead, you can use our `F.` functions like `AND(F.GT(1, 3), F.LT(1, 3))` or `SUM(F.SUB(2, 1), 2)`
+- We still do not support parentheses to change the order of operations, but you can use our `F.` functions. So for example instead of `2*(1+1)`, you should use `F.MUL(2, F.ADD(1, 1))`
 
 ## Contributing
 
@@ -61,11 +65,16 @@ use pest_derive::Parser;
 struct FormulaInner;
 
 #[derive(Debug)]
+/// `Formula`, is the main struct and entry point of this library.
 pub struct Formula<'a> {
     pairs: Pair<'a, Rule>,
 }
 
 #[derive(Debug, PartialEq)]
+/// `Expr` is the result of parsing a formula.
+///
+/// There is a difference between Excel and this library here.
+/// We don't have a `#N/A`, `#VALUE!`, `#DIV/0!`, `#NUM!`, `#NULL!` error types, instead it will return `Expr::Null`.
 pub enum Expr {
     Date(NaiveDate),
     Datetime(DateTime<Utc>),
@@ -73,27 +82,62 @@ pub enum Expr {
     Number(f64),
     String(String),
     Bool(bool),
+    Null,
 }
 
 impl<'a> Formula<'a> {
+    /// To interpret and prepare a new formula, you need to call the `new` method, like the code below:
+    ///
+    /// ```rust
+    /// use formula::{Formula, Expr, error::Error};
+    /// use anyhow::Result;
+    ///
+    /// fn main() -> Result<()> {
+    ///     let formula = Formula::new("=UPPER(TRIM('   Hello '))")?;
+    ///     let value = formula.parse()?;
+    ///     assert_eq!(value, Expr::String("HELLO".to_string()));
+    ///     Ok(())
+    /// }
+    /// ```
+    ///
+    /// # Errors
+    ///
+    /// Will return `Err` if the formula is not valid.
     pub fn new(formula: &'a str) -> Result<Self> {
-        let pairs = FormulaInner::parse(Rule::formula, formula)?
+        let pairs = FormulaInner::parse(Rule::root, formula)?
             .next()
-            .ok_or(error::Error::Parser("No formula found"))?;
+            .ok_or_else(|| error::Error::Parser("No formula found".to_string()))?;
         Ok(Self { pairs })
     }
 
+    /// Parse a formula and return the result
     ///
-    /// # Panics
+    /// # Errors
     ///
-    /// Will panic if the function is not implemented yet.
-    ///
+    /// Will return `Err` if the formula is not valid or the functions are not implemented.
     pub fn parse(self) -> Result<Expr> {
         Self::parse_pair(self.pairs)
     }
 
+    #[allow(clippy::too_many_lines)]
     fn parse_pair(pair: Pair<Rule>) -> Result<Expr> {
+        #[allow(clippy::match_same_arms)]
         let res = match pair.as_rule() {
+            // Operators
+            Rule::add => Self::parse_add(pair)?,
+            Rule::sub => Self::parse_sub(pair)?,
+            Rule::mul => Self::parse_mul(pair)?,
+            Rule::div => Self::parse_div(pair)?,
+            Rule::pow => Self::parse_pow(pair)?,
+            Rule::eq => Self::parse_eq(pair)?,
+            Rule::ne => Self::parse_ne(pair)?,
+            Rule::gt => Self::parse_gt(pair)?,
+            Rule::lt => Self::parse_lt(pair)?,
+            Rule::gte => Self::parse_gte(pair)?,
+            Rule::lte => Self::parse_lte(pair)?,
+            Rule::percent => Self::parse_percent(pair)?,
+            Rule::negate => Self::parse_negate(pair)?,
+
             // Date and time functions
             Rule::date => Self::parse_date(pair)?,
             Rule::time => Self::parse_time(pair)?,
@@ -120,7 +164,8 @@ impl<'a> Formula<'a> {
             | Rule::workdaysintl
             | Rule::yearfrac
             | Rule::days360
-            | Rule::datediff => todo!("Not implemented: {:?}", pair.as_rule()),
+            | Rule::datediff => return Err(error::Error::NotImplemented(format!("{:?}", pair.as_rule())).into()),
+
             // Text functions
             Rule::left => Self::parse_left(pair)?,
             Rule::leftb => Self::parse_leftb(pair)?,
@@ -164,13 +209,34 @@ impl<'a> Formula<'a> {
             | Rule::textsplit
             | Rule::value
             | Rule::valuetotext
-            | Rule::bahttext => todo!("Not implemented: {:?}", pair.as_rule()),
+            | Rule::bahttext => return Err(error::Error::NotImplemented(format!("{:?}", pair.as_rule())).into()),
             // Engineering functions
             // Financial functions
             // Logical functions
+            Rule::and => Self::parse_and(pair)?,
+            Rule::or => Self::parse_or(pair)?,
+            Rule::xor => Self::parse_xor(pair)?,
+            Rule::not => Self::parse_not(pair)?,
+            Rule::if_ => Self::parse_if(pair)?,
+            Rule::ifna => Self::parse_ifna(pair)?,
+            Rule::iferror => Self::parse_iferror(pair)?,
+            // TODO remaining text functions
+            Rule::let_
+            | Rule::bycol
+            | Rule::byrow
+            | Rule::makearray
+            | Rule::reduce
+            | Rule::scan
+            | Rule::map
+            | Rule::lambda
+            | Rule::switch
+            | Rule::ifs => return Err(error::Error::NotImplemented(format!("{:?}", pair.as_rule())).into()),
             // Math functions
             // Statistical functions
             // Web functions
+            Rule::encodeurl => Self::parse_encodeurl(pair)?,
+            Rule::filterxml => Self::parse_filterxml(pair)?,
+            Rule::webservice => Self::parse_webservice(pair)?,
 
             // Basic types
             Rule::num => {
@@ -184,6 +250,7 @@ impl<'a> Formula<'a> {
             Rule::bool_true => Expr::Bool(true),
             Rule::bool_false => Expr::Bool(false),
             Rule::formula
+            | Rule::root
             | Rule::OP
             | Rule::F
             | Rule::CF
@@ -193,13 +260,42 @@ impl<'a> Formula<'a> {
             | Rule::inner
             | Rule::char_
             | Rule::basic_types
+            | Rule::operators
             | Rule::datetime_functions
             | Rule::text_functions
+            | Rule::logical_functions
+            | Rule::web_functions
             | Rule::WHITESPACE => {
                 unreachable!()
             }
         };
 
         Ok(res)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::Formula;
+
+    #[test]
+    #[allow(clippy::too_many_lines)]
+    fn test_parse_basic_types() {
+        let formula = Formula::new("=TRUE").unwrap();
+        let value = formula.parse().unwrap();
+        assert_eq!(value, Expr::Bool(true));
+
+        let formula = Formula::new("=TRUE()").unwrap();
+        let value = formula.parse().unwrap();
+        assert_eq!(value, Expr::Bool(true));
+
+        let formula = Formula::new("=25").unwrap();
+        let value = formula.parse().unwrap();
+        assert_eq!(value, Expr::Number(25.0));
+
+        let formula = Formula::new("='TEST'").unwrap();
+        let value = formula.parse().unwrap();
+        assert_eq!(value, Expr::String("TEST".to_string()));
     }
 }