+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased (2022-09-xx)
+
+### Added
+
+- Support of basic MS Excel logical functions
+- Support of most MS Excel web functions
+
 ## 0.0.1 (2022-09-05)
 
 ### Added